Move SpectrumEvaluator to gammapy.spectrum.core
