+++ gammapy/spectrum/dataset.py
@@ -12,8 +12,7 @@ from gammapy.utils.fits import energy_axis_to_ebounds
 from gammapy.utils.fitting import Dataset, Parameters
 from gammapy.utils.random import get_random_state
 from gammapy.utils.scripts import make_path
-from .core import CountsSpectrum
-from .utils import SpectrumEvaluator
+from .core import CountsSpectrum, SpectrumEvaluator
 
 __all__ = ["SpectrumDatasetOnOff", "SpectrumDataset", "SpectrumDatasetOnOffStacker"]
